+++ src/main/java/cofh/redstonearsenal/item/tool/ItemAxeRF.java
@@ -1,4 +1,4 @@
-package redstonearsenal.item.tool;
+package cofh.redstonearsenal.item.tool;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
